+++ Source/Scene/ImageryLayerCollection.js
@@ -348,7 +348,7 @@ define([
      *
      * @example
      * var pickRay = viewer.camera.getPickRay(windowPosition);
-     * var featuresPromise = viewer.scene.imageryLayers.pickImageryLayerFeatures(pickRay, viewer.scene);
+     * var featuresPromise = viewer.imageryLayers.pickImageryLayerFeatures(pickRay, viewer.scene);
      * if (!Cesium.defined(featuresPromise)) {
      *     console.log('No features picked.');
      * } else {